fixed mapping and worked a bit on NamerServerDatabase

+++ src/test/java/nintendods/ds_project/helper/MappingTests.java
@@ -8,41 +8,41 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 public class MappingTests {
     @Test
     public void checkMappingUpperLimit(){
-        long oldMin = 50;
-        long oldMax = 100;
+        int oldMin = 50;
+        int oldMax = 100;
 
-        long newMin = 0;
-        long newMax = 10;
+        int newMin = 0;
+        int newMax = 10;
 
-        long checkValue = Mapping.map(100, oldMin, oldMax, newMin, newMax);
-        long actualValue = 10;
+        int checkValue = Mapping.map(100, oldMin, oldMax, newMin, newMax);
+        int actualValue = 10;
 
         assertEquals(actualValue, checkValue);
     }
 
     @Test
     public void checkMappingLowerLimit(){
-        long oldMin = 50;
-        long oldMax = 100;
+        int oldMin = 50;
+        int oldMax = 100;
 
-        long newMin = 0;
-        long newMax = 10;
+        int newMin = 0;
+        int newMax = 10;
 
-        long checkValue = Mapping.map(50, oldMin, oldMax, newMin, newMax);
-        long actualValue = 0;
+        int checkValue = Mapping.map(50, oldMin, oldMax, newMin, newMax);
+        int actualValue = 0;
 
         assertEquals(actualValue, checkValue);
     }
     @Test
     public void checkMappingOutOfBoundNumbers(){
-        long oldMin = 50;
-        long oldMax = 100;
+        int oldMin = 50;
+        int oldMax = 100;
 
-        long newMin = 0;
-        long newMax = 10;
+        int newMin = 0;
+        int newMax = 10;
 
-        long checkValue = Mapping.map(0, oldMin, oldMax, newMin, newMax);
-        long actualValue = newMin;
+        int checkValue = Mapping.map(0, oldMin, oldMax, newMin, newMax);
+        int actualValue = newMin;
 
         assertEquals(actualValue, checkValue);
 
@@ -54,14 +54,14 @@ public class MappingTests {
 
     @Test
     public void checkMappingZero(){
-        long oldMin = 0;
-        long oldMax = 0;
+        int oldMin = 0;
+        int oldMax = 0;
 
-        long newMin = 0;
-        long newMax = 0;
+        int newMin = 0;
+        int newMax = 0;
 
-        long checkValue = Mapping.map(0, oldMin, oldMax, newMin, newMax);
-        long actualValue = newMin;
+        int checkValue = Mapping.map(0, oldMin, oldMax, newMin, newMax);
+        int actualValue = newMin;
 
         assertEquals(actualValue, checkValue);
     }
